python/flask-healthz-db/requirements.txt
click==8.1.2
Flask==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.1.1
gunicorn==20.1.0
mysql.connector
